BUG: Fix osna idxmax over labels without time
With removing pandas version <=v.1.2.5 we got also a change that broke
the osna method in location identification.
I could not really trace the origin of the change but in the end the
`idxmax` from pandas uses `nanargmax` from numpy. That works even if
everything is nan, but only if it is nan and not any other NA value.
In our case that was pd.NaT for the Timedeltas. In my opinion that is a bug
in pandas as their documentation is stating that it works for NA
(that includes NaT).
Well in the end I only had to change the dtype of the aggregated column. :D

TLDR: Changed dtype such that idxmax gets np.nan